Test: 실패하는 테스트들 모두 수정
